requirements.txt
absl-py==0.7.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
gast==0.2.2
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-bigquery==1.13.0
google-cloud-core==1.0.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.21.1
h5py==2.9.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
mock==3.0.5
nose==1.3.7
numpy==1.16.3
pandas==0.24.2
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
rsa==4.0
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
TBB==0.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.4
